+++ app/javascript/src/components/editor/FindReplaceAll.svelte
@@ -1,6 +1,7 @@
 <script>
   import { items, currentItem, editorStates } from "../../stores/editor"
-  import { getItemById, replaceBetween, setCurrentItemById, updateItem } from "../../utils/editor"
+  import { getItemById, setCurrentItemById, updateItem } from "../../utils/editor"
+  import { replaceBetween } from "../../utils/parse"
   import { escapeable } from "../actions/escapeable"
   import { fade, fly } from "svelte/transition"
   import { tick } from "svelte"
