+++ app/javascript/src/utils/setCssVariable.js
@@ -1,0 +1,3 @@
+export function setCssVariable(key, value) {
+  document.body.style.setProperty(`--${ key }`, value)
+}
